+++ src/nagios_linux_test.go
@@ -1,9 +1,13 @@
 // +build linux
+
 package main
 
 import (
+	"os"
+	"sync"
 	"testing"
 
+	"github.com/newrelic/infra-integrations-sdk/integration"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -19,4 +23,42 @@ func Test_runCommand_returns1(t *testing.T) {
 	assert.Equal(t, 2, exit)
 	assert.Empty(t, stdout)
 	assert.Empty(t, stderr)
+}
+
+func Test_collectServiceCheck(t *testing.T) {
+	i, _ := integration.New("test", "test")
+	sc := serviceCheck{
+		Name:        "testname",
+		Command:     []string{"echo", "testout"},
+		Labels:      map[string]string{"testkey": "testval"},
+		ParseOutput: false,
+	}
+	serverName, err := os.Hostname()
+	if err != nil {
+		panic(err)
+	}
+
+	expectedMetrics := map[string]interface{}{
+		"serviceCheck.name":    "testname",
+		"serviceCheck.status":  float64(0),
+		"serviceCheck.message": "testout\n",
+		"serviceCheck.error":   "",
+		"serviceCheck.command": "echo testout",
+		"serverName":           serverName,
+		"displayName":          "testname",
+		"entityName":           "serviceCheck:testname",
+		"event_type":           "NagiosServiceCheckSample",
+		"testkey":              "testval",
+	}
+
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go collectServiceCheck(sc, i, &wg, "NagiosServiceCheckSample")
+	wg.Wait()
+
+	id := integration.NewIDAttribute("executing_host", serverName)
+	e, _ := i.Entity("testname", "serviceCheck", id)
+	metrics := e.Metrics[0].Metrics
+
+	assert.Equal(t, expectedMetrics, metrics)
 }
