+++ src/nagios_test.go
@@ -51,44 +51,6 @@ func Test_parseConfigFile_UnrestrictiveError(t *testing.T) {
 	assert.Nil(t, config)
 }
 
-func Test_collectServiceCheck(t *testing.T) {
-	i, _ := integration.New("test", "test")
-	sc := serviceCheck{
-		Name:        "testname",
-		Command:     []string{"echo", "testout"},
-		Labels:      map[string]string{"testkey": "testval"},
-		ParseOutput: false,
-	}
-	serverName, err := os.Hostname()
-	if err != nil {
-		panic(err)
-	}
-
-	expectedMetrics := map[string]interface{}{
-		"serviceCheck.name":    "testname",
-		"serviceCheck.status":  float64(0),
-		"serviceCheck.message": "testout\n",
-		"serviceCheck.error":   "",
-		"serviceCheck.command": "echo testout",
-		"serverName":           serverName,
-		"displayName":          "testname",
-		"entityName":           "serviceCheck:testname",
-		"event_type":           "NagiosServiceCheckSample",
-		"testkey":              "testval",
-	}
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-	go collectServiceCheck(sc, i, &wg, "NagiosServiceCheckSample")
-	wg.Wait()
-
-	id := integration.NewIDAttribute("executing_host", serverName)
-	e, _ := i.Entity("testname", "serviceCheck", id)
-	metrics := e.Metrics[0].Metrics
-
-	assert.Equal(t, expectedMetrics, metrics)
-}
-
 func Test_collectServiceCheck_InvalidNameError(t *testing.T) {
 	i, _ := integration.New("test", "test")
 	sc := serviceCheck{
